jenks.py
This is an implementation of the PySAL package that provides a function for calculating Jenks Natural Breaks categories.  The script assumes that the IQHO data provided by JLC Géomatique is the basis for the calculation of Jenks categories, and that the number of categories required are four in number (corresponding to differing levels of suitability for moose hunting, from very good to very bad).

At present the documentation is in English but a French version will likely follow.